Improve const correctness, fix string comparisons

+++ src/algorithms/signal_source/adapters/two_bit_packed_file_signal_source.h
@@ -129,7 +129,7 @@ public:
     }
 
 private:
-    unsigned long long samples_;
+    uint64_t samples_;
     long sampling_frequency_;
     std::string filename_;
     std::string item_type_;
